pyproject.toml
[project]
name = "quick-match"
version = "0.1.0"
description = ""
authors = [
    {name = "PlasmOwO",email = "thomas.plasmo@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "streamlit (>=1.52.2,<2.0.0)"
]

[tool.poetry]
packages = [{include = "quick_match", from = "src"}]
include =[
    "src/quick_match/webapp/style.css"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
app = "quick_match.streamlit_app_runner:main"
